Revert "Routing implementation"
