--- src/components/ArtistsCard.jsx
@@ -1,18 +1,0 @@
-import "../styles/components/joinTheResistance.css";
-
-
-function ArtistsCard() {
-
-    return (
-
-        <div className="card">
-            <h2>🎤 Artists</h2>
-            <p>
-                Connect directly with fan base, build your following, reward
-                loyalty, protect your IP, royalties, and Base from excessive fees.
-            </p>
-        </div>
-    )
-}
-
-export default ArtistsCard;
