--- src/components/FansCard.jsx
@@ -1,18 +1,0 @@
-import "../styles/components/joinTheResistance.css";
-
-
-function FansCard() {
-
-    return (
-        <div className="card">
-            <h2>🎟️ Fans</h2>
-            <p>
-                Secure your ticket, claim exclusive rewards, and build the
-                underground movement that brings Fans, Artists and Venues together
-                to overthrow the 🎟️ "Master."
-            </p>
-        </div>
-    )
-}
-
-export default FansCard;
