--- src/components/HomePage.jsx
@@ -1,9 +1,0 @@
-
-
-export default function HomePage() {
-    return (
-        <>
-            <h1>HomePage</h1>
-        </>
-    )
-}
