--- src/components/VenuesCard.jsx
@@ -1,18 +1,0 @@
-import "../styles/components/joinTheResistance.css";
-
-function VenuesCard() {
-
-    return (
-        <div className="card">
-            <h2>🏟️ Venues</h2>
-            <p>
-                Next generation ticketing now! Smart contracts prevent fraud,
-                control resale, and build community trust around your venue while we
-                handle your ticketing with enhanced sales analytics and no bank fees
-                on USDC stablecoin transactions.
-            </p>
-        </div>
-    )
-}
-
-export default VenuesCard;
