+++ src/main.jsx
@@ -4,16 +4,10 @@ import { createRoot } from "react-dom/client";
 import { BrowserRouter as Router } from "react-router-dom";
 
 import App from "./App.jsx";
-import APIProvider from "./components/ApiFetch.jsx";
 import "./styles/index.css";
 
 createRoot(document.getElementById("root")).render(
-  <StrictMode>
-    <APIProvider>
-      <Router>
-        <App />
-      </Router>
-    </APIProvider>
-  </StrictMode>
-
+  <Router>
+    <App />
+  </Router>
 );
